runpod/rembg_requirements.txt
runpod>=1.6.0
rembg[gpu]>=2.0.50
Pillow>=10.0.0
onnxruntime-gpu>=1.16.0
